+++ packages/api/amplify_api_dart/lib/src/graphql/web_socket/connection/amplify_web_socket.dart
@@ -1,25 +1,31 @@
 import 'package:amplify_api_dart/src/decorators/web_socket_auth_utils.dart';
+import 'package:amplify_api_dart/src/graphql/web_socket/connection/amplify_web_socket_connection.dart'
+    if (dart.library.io) 'amplify_web_socket_io.dart'
+    if (dart.library.html) 'amplify_web_socket_web.dart';
 import 'package:amplify_core/amplify_core.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
 
-/// An Abstract class that provides a way to connect to a WebSocket server.
-/// Provides platform specific implementations.
-class AmplifyWebSocketConnection {
-  /// Constructor for AmplifyWebSocket
+/// Amplify Web socket connection class that provides a way to connect to a WebSocket server.
+class AmplifyWebSocket {
+  /// Constructor for AmplifyWebSocketConnection
   /// Takes in an [AWSApiConfig] and an [AmplifyAuthProviderRepository]
-  AmplifyWebSocketConnection({
+  AmplifyWebSocket({
     required this.config,
     required this.authProviderRepository,
   });
 
-  /// Connect to a web socket server at the specified [uri].
-  Future<WebSocketChannel> connect(Uri uri) async {
-    throw UnimplementedError('connect() has not been implemented.');
+  /// The [Uri] for the web socket
+  Future<Uri> get uri async {
+    return generateConnectionUri(config, authProviderRepository);
   }
 
-  /// fetch auth headers
-  Future<Map<String, String>> getAuthHeaders() async {
-    return authorizeConnection(config, authProviderRepository);
+  /// Connect to a web socket server.
+  Future<WebSocketChannel> connect() async {
+    final uri = await this.uri;
+    return AmplifyWebSocketConnection(
+      config: config,
+      authProviderRepository: authProviderRepository,
+    ).connect(uri);
   }
 
   /// AWS Config
@@ -27,7 +33,4 @@ class AmplifyWebSocketConnection {
 
   /// Amplify Auth Provider
   final AmplifyAuthProviderRepository authProviderRepository;
-
-  /// Websocket Protocols
-  List<String> webSocketProtocols = ['graphql-ws'];
 }
