+++ packages/api/amplify_api_dart/lib/src/graphql/web_socket/connection/amplify_web_socket_connection.dart
@@ -1,31 +1,25 @@
 import 'package:amplify_api_dart/src/decorators/web_socket_auth_utils.dart';
-import 'package:amplify_api_dart/src/graphql/web_socket/connection/amplify_web_socket.dart'
-    if (dart.library.io) 'amplify_web_socket_io.dart'
-    if (dart.library.html) 'amplify_web_socket_web.dart';
 import 'package:amplify_core/amplify_core.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
 
-/// Amplify Web socket connection class that provides a way to connect to a WebSocket server.
-class AmplifyWebSocket {
-  /// Constructor for AmplifyWebSocketConnection
+/// An Abstract class that provides a way to connect to a WebSocket server.
+/// Provides platform specific implementations.
+class AmplifyWebSocketConnection {
+  /// Constructor for AmplifyWebSocket
   /// Takes in an [AWSApiConfig] and an [AmplifyAuthProviderRepository]
-  AmplifyWebSocket({
+  AmplifyWebSocketConnection({
     required this.config,
     required this.authProviderRepository,
   });
 
-  /// The [Uri] for the web socket
-  Future<Uri> get uri async {
-    return generateConnectionUri(config, authProviderRepository);
+  /// Connect to a web socket server at the specified [uri].
+  Future<WebSocketChannel> connect(Uri uri) async {
+    throw UnimplementedError('connect() has not been implemented.');
   }
 
-  /// Connect to a web socket server.
-  Future<WebSocketChannel> connect() async {
-    final uri = await this.uri;
-    return AmplifyWebSocketConnection(
-      config: config,
-      authProviderRepository: authProviderRepository,
-    ).connect(uri);
+  /// fetch auth headers
+  Future<Map<String, String>> getAuthHeaders() async {
+    return authorizeConnection(config, authProviderRepository);
   }
 
   /// AWS Config
@@ -33,4 +27,7 @@ class AmplifyWebSocket {
 
   /// Amplify Auth Provider
   final AmplifyAuthProviderRepository authProviderRepository;
+
+  /// Websocket Protocols
+  List<String> webSocketProtocols = ['graphql-ws'];
 }
